Separate power state read and write nodes
