+++ package.json
@@ -30,7 +30,8 @@
       "ewelink-devices": "src/devices/devices.js",
       "ewelink-generic-device": "src/generic-device/generic-device.js",
       "ewelink-firmware-version": "src/firmware-version/firmware-version.js",
-      "ewelink-power-state": "src/power-state/power-state.js",
+      "ewelink-power-state-read": "src/power-state-read/power-state-read.js",
+      "ewelink-power-state-write": "src/power-state-write/power-state-write.js",
       "ewelink-channel-count": "src/channel-count/channel-count.js",
       "ewelink-event-listener": "src/event-listener/event-listener.js",
       "ewelink-power-usage": "src/power-usage/power-usage.js",
